Separate `PlayerDamageByPlayer` handler and add methods
