+++ api/src/main/java/xyz/srnyx/annoyingapi/events/PlayerDamageByPlayerEvent.java
@@ -1,17 +1,12 @@
 package xyz.srnyx.annoyingapi.events;
 
-import org.bukkit.Bukkit;
-import org.bukkit.entity.Entity;
+import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
 import org.bukkit.event.HandlerList;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
-
-import xyz.srnyx.annoyingapi.AnnoyingListener;
-import xyz.srnyx.annoyingapi.AnnoyingPlugin;
 
 
 public class PlayerDamageByPlayerEvent extends EntityDamageByEntityEvent {
@@ -21,37 +16,33 @@ public class PlayerDamageByPlayerEvent extends EntityDamageByEntityEvent {
         super(damager, damagee, cause, damage);
     }
 
-    @Contract(pure = true)
+    @NotNull @Contract(pure = true)
     public static HandlerList getHandlerList() {
         return HANDLERS;
     }
 
-    @Override
+    @Override @NotNull
     public HandlerList getHandlers() {
         return HANDLERS;
     }
 
-    public static class EntityDamageByEntityListener implements AnnoyingListener {
-        @NotNull private final AnnoyingPlugin plugin;
+    @Override @NotNull
+    public Player getDamager() {
+        return (Player) super.getDamager();
+    }
 
-        @Contract(pure = true)
-        public EntityDamageByEntityListener(@NotNull AnnoyingPlugin plugin) {
-            this.plugin = plugin;
-        }
+    @Override @NotNull
+    public Player getEntity() {
+        return (Player) super.getEntity();
+    }
 
-        @NotNull
-        public AnnoyingPlugin getPlugin() {
-            return plugin;
-        }
+    @Override @NotNull
+    public EntityType getEntityType() {
+        return EntityType.PLAYER;
+    }
 
-        @EventHandler
-        public void onEntityDamageByEntity(@NotNull EntityDamageByEntityEvent event) {
-            final Entity damager = event.getDamager();
-            final Entity damagee = event.getEntity();
-            if (!(damager instanceof Player) || !(damagee instanceof Player)) return;
-            final PlayerDamageByPlayerEvent newEvent = new PlayerDamageByPlayerEvent((Player) damager, (Player) damagee, event.getCause(), event.getDamage());
-            Bukkit.getPluginManager().callEvent(newEvent);
-            event.setCancelled(newEvent.isCancelled());
-        }
+    @NotNull
+    public Player getDamagee() {
+        return getEntity();
     }
 }
